D3ASIM-3563: Fixed bug in tests.
